refactor: Poll Details -> Proposal Details rename
