+++ app/src/components/ProposalDetails.vue
